added Line class and algorithm to find line intersect.

+++ CPP_Course/ex2/Point.cpp
@@ -16,7 +16,7 @@ Point::Point() : _x(0), _y(0)
 {
 }
 
-Point::Point(int x, int y): _x(x), _y(y)
+Point::Point(double x, double y): _x(x), _y(y)
 {
 }
 
@@ -24,7 +24,7 @@ Point::~Point()
 {
 }
 
-void Point::set(const int x, const int y)
+void Point::set(const double x, const double y)
 {
 	_x = x;
 	_y = y;
